restructure: move nicesp into src/, compile dplib separately
This puts everything into a single source tree and compiles
different things from the same source tree. The client-specific
code is in src/aocmultiny, the DirectPlay library is in src/dplib,
the libnice Service Provider is in src/nicesp, and random other
fluff is simply in src/.

Currently dependencies between these are managed manually in the
Makefile but I don't expect many more targets being added, so
that's probably fine.
