+++ src/aocmultiny/irc/Channel.cpp
