+++ src/aocmultiny/irc/Channel.hpp
