+++ src/dplib/DPLConnection.cpp
@@ -5,7 +5,6 @@
 #include "DPName.hpp"
 #include "DPLConnection.hpp"
 
-namespace aocmultiny {
 namespace dplib {
 
 const GUID DPSPGUID_NICE = { 0xe2dd8ebe, 0x1f03, 0x43b7, { 0x8d, 0x92, 0x9c, 0x6c, 0x2f, 0x5c, 0x44, 0x26 } };
@@ -40,5 +39,4 @@ DPLConnection::~DPLConnection () {
   delete this->dpConnection;
 }
 
-}
 }
