+++ src/dplib/util.cpp
@@ -1,7 +1,6 @@
 #include <string>
 #include <dplay.h>
 
-namespace aocmultiny {
 namespace dplib {
 
 std::string getDPError (HRESULT hr) {
@@ -61,5 +60,4 @@ std::string getDPError (HRESULT hr) {
   }
 }
 
-}
 }
