+++ src/dplib/util.hpp
@@ -2,10 +2,8 @@
 #include <string>
 #include <dplay.h>
 
-namespace aocmultiny {
 namespace dplib {
 
 std::string getDPError (HRESULT hr);
 
-}
 }
