+++ src/nicesp/NiceAgent.cpp
