+++ src/nicesp/NiceAgent.hpp
